Update BB field type to uint16_t as per protocol change
Make a different routine to print the BB field
which needs to be formatted accurately:
e.g. 555 -> 55.5
This is due to the way brake bias is stored and sent over the wire.
TODO: fix handling 0.

+++ Device/UnoR3/arduino-ide/simdisplay/SimDisplayProtocol.h
@@ -31,7 +31,7 @@ struct SimDisplayPacket {
 	uint8_t tcaction;
 	uint8_t abs;
 	uint8_t absaction;
-	uint8_t bb;
+	uint16_t bb;
 	uint8_t remlaps;
 	uint8_t map;
 	uint8_t airt;
